# Improve performance and reduce allocations of TemplateSegment