--- a/benchmarks/Microsoft.AspNetCore.Routing.Performance/Template/TemplateSegmentBenchmark.cs
+++ b/benchmarks/Microsoft.AspNetCore.Routing.Performance/Template/TemplateSegmentBenchmark.cs
@@ -1,6 +1,9 @@
-﻿using BenchmarkDotNet.Attributes;
-using Microsoft.AspNetCore.Routing.Patterns;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Routing.Patterns;
+using BenchmarkDotNet.Attributes;
 
 namespace Microsoft.AspNetCore.Routing.Template
 {