Improves performance and reduce allocations of **TemplateSegment(RoutePatternPathSegment other)** ctor, and adds a few unit tests.

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17134
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3914069 Hz, Resolution=255.4886 ns, Timer=TSC
.NET Core SDK=2.1.400
  [Host]     : .NET Core 2.2.0-preview3-26927-02 (CoreCLR 4.6.26927.02, CoreFX 4.6.26927.02), 64bit RyuJIT
  Job-EFQGFJ : .NET Core 2.2.0-preview3-26927-02 (CoreCLR 4.6.26927.02, CoreFX 4.6.26927.02), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.2  
RunStrategy=Throughput  
```

**Before:**

|                       Method |     Mean |     Error |    StdDev |        Op/s |  Gen 0 | Allocated |
|----------------------------- |---------:|----------:|----------:|------------:|-------:|----------:|
| Ctor_RoutePatternPathSegment | 231.5 ns | 0.7288 ns | 0.6818 ns | 4,319,262.6 | 0.0057 |     536 B |
|    ToRoutePatternPathSegment | 124.3 ns | 1.5693 ns | 1.4679 ns | 8,043,307.1 | 0.0036 |     320 B |

**After:**

|                       Method |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Allocated |
|----------------------------- |----------:|---------:|---------:|-------------:|-------:|----------:|
| Ctor_RoutePatternPathSegment |  87.38 ns | 1.747 ns | 2.449 ns | 11,444,012.9 | 0.0043 |     408 B |
|    ToRoutePatternPathSegment | 124.76 ns | 1.564 ns | 1.463 ns |  8,015,433.9 | 0.0036 |     320 B |