@drieseng 2018-10-15
This is now ready for review.
Just let me know if you want me to get rid of the benchmark.