@drieseng 2018-10-21
I've removed the benchmark. That should have addressed all PR feedback.

For reference, the benchmark is now available [here](https://gist.github.com/drieseng/56ab1f0bcd9902a3f44b4d1c9caf8d6e).